+++ tests/TestCase/View/Helper/UserHelperTest.php
@@ -11,6 +11,7 @@
 
 namespace CakeDC\Users\Test\TestCase\View\Helper;
 
+use CakeDC\Users\Model\Entity\SocialAccount;
 use CakeDC\Users\View\Helper\UserHelper;
 use Cake\Core\App;
 use Cake\Core\Configure;
@@ -29,6 +30,19 @@ use Cake\View\View;
  */
 class UserHelperTest extends TestCase
 {
+    /**
+     * Keep the original config for oauth
+     *
+     * @var array
+     */
+    private $oauthConfig;
+
+    /**
+     * Keep original config Users.Social.login
+     *
+     * @var boolean
+     */
+    private $socialLogin;
 
     /**
      * setUp method
@@ -37,6 +51,11 @@ class UserHelperTest extends TestCase
      */
     public function setUp()
     {
+        if ($this->oauthConfig === null) {
+            $this->oauthConfig = (array)Configure::read('OAuth');
+            $this->socialLogin = Configure::read('Users.Social.login');
+        }
+
         parent::setUp();
         Plugin::routes('CakeDC/Users');
         $this->View = $this->getMockBuilder('Cake\View\View')
@@ -62,6 +81,8 @@ class UserHelperTest extends TestCase
      */
     public function tearDown()
     {
+        Configure::write('OAuth', $this->oauthConfig);
+        Configure::write('Users.Social.login', $this->socialLogin);
         unset($this->User);
 
         parent::tearDown();
@@ -223,5 +244,135 @@ class UserHelperTest extends TestCase
         $result = $this->User->socialLogin('facebook');
         $this->assertEquals('<a href="/auth/facebook" class="btn btn-social btn-facebook"><i class="fa fa-facebook"></i>Iniciar sesión con Facebook</a>', $result);
         I18n::locale('en_US');
+    }
+
+    /**
+     * Test social connect link list
+     *
+     * @return void
+     */
+    public function testSocialConnectLinkList()
+    {
+        Configure::write('Users.Social.login', true);
+
+        Configure::write('OAuth.providers.facebook.options.clientId', 'testclientidtestclientid');
+        Configure::write('OAuth.providers.facebook.options.clientSecret', 'testclientsecrettestclientsecret');
+
+        Configure::write('OAuth.providers.google.options.clientId', 'testclientidgoogtestclientidgoog');
+        Configure::write('OAuth.providers.google.options.clientSecret', 'testclientsecretgoogtestclientsecretgoog');
+
+        $actual = $this->User->socialConnectLinkList();
+        $expected = '<a href="/link-social/facebook" class="btn btn-social btn-facebook"><span class="fa fa-facebook"></span> Connect with Facebook</a>';
+        $expected .= '<a href="/link-social/google" class="btn btn-social btn-google"><span class="fa fa-google"></span> Connect with Google</a>';
+        $this->assertEquals($expected, $actual);
+    }
+
+    /**
+     * Test social connect link list, user is connected with facebook
+     *
+     * @return void
+     */
+    public function testSocialConnectLinkListIsConnectedWithFacebook()
+    {
+        Configure::write('Users.Social.login', true);
+
+        Configure::write('OAuth.providers.facebook.options.clientId', 'testclientidtestclientid');
+        Configure::write('OAuth.providers.facebook.options.clientSecret', 'testclientsecrettestclientsecret');
+
+        Configure::write('OAuth.providers.google.options.clientId', 'testclientidgoogtestclientidgoog');
+        Configure::write('OAuth.providers.google.options.clientSecret', 'testclientsecretgoogtestclientsecretgoog');
+
+        $socialAccounts = [
+            new SocialAccount([
+                'id' => '00000000-0000-0000-0000-000000000001',
+                'user_id' => '00000000-0000-0000-0000-000000000001',
+                'provider' => 'Facebook',
+                'username' => 'user-1-fb',
+                'reference' => 'reference-1-1234',
+                'avatar' => 'Lorem ipsum dolor sit amet',
+                'description' => 'Lorem ipsum dolor sit amet, aliquet feugiat. Convallis morbi fringilla gravida, phasellus feugiat dapibus velit nunc, pulvinar eget sollicitudin venenatis cum nullam, vivamus ut a sed, mollitia lectus. Nulla vestibulum massa neque ut et, id hendrerit sit, feugiat in taciti enim proin nibh, tempor dignissim, rhoncus duis vestibulum nunc mattis convallis.',
+                'token' => 'token-1234',
+                'token_secret' => 'Lorem ipsum dolor sit amet',
+                'token_expires' => '2015-05-22 21:52:44',
+                'active' => false,
+                'data' => '',
+                'created' => '2015-05-22 21:52:44',
+                'modified' => '2015-05-22 21:52:44'
+            ])
+        ];
+        $actual = $this->User->socialConnectLinkList($socialAccounts);
+        $expected = '<a class="btn btn-social btn-facebook disabled"><span class="fa fa-facebook"></span> Connected with Facebook</a>';
+        $expected .= '<a href="/link-social/google" class="btn btn-social btn-google"><span class="fa fa-google"></span> Connect with Google</a>';
+        $this->assertEquals($expected, $actual);
+    }
+
+    /**
+     * Test social connect link list, social is not enabled
+     *
+     * @return void
+     */
+    public function testSocialConnectLinkListSocialIsNotEnabled()
+    {
+        Configure::write('Users.Social.login', false);
+
+        Configure::write('OAuth.providers.facebook.options.clientId', 'testclientidtestclientid');
+        Configure::write('OAuth.providers.facebook.options.clientSecret', 'testclientsecrettestclientsecret');
+
+        Configure::write('OAuth.providers.google.options.clientId', 'testclientidgoogtestclientidgoog');
+        Configure::write('OAuth.providers.google.options.clientSecret', 'testclientsecretgoogtestclientsecretgoog');
+
+        $socialAccounts = [
+            new SocialAccount([
+                'id' => '00000000-0000-0000-0000-000000000001',
+                'user_id' => '00000000-0000-0000-0000-000000000001',
+                'provider' => 'Facebook',
+                'username' => 'user-1-fb',
+                'reference' => 'reference-1-1234',
+                'avatar' => 'Lorem ipsum dolor sit amet',
+                'description' => 'Lorem ipsum dolor sit amet, aliquet feugiat. Convallis morbi fringilla gravida, phasellus feugiat dapibus velit nunc, pulvinar eget sollicitudin venenatis cum nullam, vivamus ut a sed, mollitia lectus. Nulla vestibulum massa neque ut et, id hendrerit sit, feugiat in taciti enim proin nibh, tempor dignissim, rhoncus duis vestibulum nunc mattis convallis.',
+                'token' => 'token-1234',
+                'token_secret' => 'Lorem ipsum dolor sit amet',
+                'token_expires' => '2015-05-22 21:52:44',
+                'active' => false,
+                'data' => '',
+                'created' => '2015-05-22 21:52:44',
+                'modified' => '2015-05-22 21:52:44'
+            ])
+        ];
+        $actual = $this->User->socialConnectLinkList($socialAccounts);
+        $expected = '';
+        $this->assertEquals($expected, $actual);
+    }
+
+    /**
+     * Test social connect link list, social is enabled but any provider was configured
+     *
+     * @return void
+     */
+    public function testSocialConnectLinkListSocialEnabledButNotConfiguredProvider()
+    {
+        Configure::write('Users.Social.login', true);
+
+        $socialAccounts = [
+            new SocialAccount([
+                'id' => '00000000-0000-0000-0000-000000000001',
+                'user_id' => '00000000-0000-0000-0000-000000000001',
+                'provider' => 'Facebook',
+                'username' => 'user-1-fb',
+                'reference' => 'reference-1-1234',
+                'avatar' => 'Lorem ipsum dolor sit amet',
+                'description' => 'Lorem ipsum dolor sit amet, aliquet feugiat. Convallis morbi fringilla gravida, phasellus feugiat dapibus velit nunc, pulvinar eget sollicitudin venenatis cum nullam, vivamus ut a sed, mollitia lectus. Nulla vestibulum massa neque ut et, id hendrerit sit, feugiat in taciti enim proin nibh, tempor dignissim, rhoncus duis vestibulum nunc mattis convallis.',
+                'token' => 'token-1234',
+                'token_secret' => 'Lorem ipsum dolor sit amet',
+                'token_expires' => '2015-05-22 21:52:44',
+                'active' => false,
+                'data' => '',
+                'created' => '2015-05-22 21:52:44',
+                'modified' => '2015-05-22 21:52:44'
+            ])
+        ];
+        $actual = $this->User->socialConnectLinkList($socialAccounts);
+        $expected = '';
+        $this->assertEquals($expected, $actual);
     }
 }